revert: refactor: emit less events (a0e817998b085df4f65f0775c7a9bb87af36e7fb)
in the case of
copy (local) -> recieve (telltail) -> restore
if we've originally copied a format other than text the restoration would
give us empty item in our clipboard. This will happen because we are only handling texts
and the clipboard api we are using will give us empty string in case of non-text copy.